fix(LevelGenerator): initial width and height are now ints
To prevent unexpected behaviour

+++ Assets/Scripts/LevelGeneration/LevelGenerator.cs
@@ -6,9 +6,9 @@ using UnityEngine;
 public class LevelGenerator : MonoBehaviour
 {
     [SerializeField]
-    private float _initialWidth = 50;
+    private int _initialWidth = 50;
     [SerializeField]
-    private float _initialHeight = 50;
+    private int _initialHeight = 50;
     [SerializeField]
     private Vector2 _initialCenter = new Vector2(0, 0);
     [SerializeField]
